project2_image_recognition/requirements.txt
opencv-python
scikit-image
scikit-learn
numpy
scipy
matplotlib
torch>=2.0
torchvision
tqdm
Pillow>=9.5.0
pandas
python-dateutil>=2.8.2
seaborn